[ 1922458 ] [Core] multiple project executables
darcs-hash:20080321203329-34f1c-21577c2d1e75e481b8efa42fe5fca986c64c6b54.gz

+++ net.sf.eclipsefp.haskell.core/src/net/sf/eclipsefp/haskell/core/builder/CleanOutFoldersOperation.java
@@ -56,9 +56,11 @@ class CleanOutFoldersOperation implements IWorkspaceRunnable {
 
   private void deleteExe( final IProgressMonitor mon ) throws CoreException {
     mon.subTask( CoreTexts.cleanOutFoldersOperation_removingExes );
-    IFile file = ResourceUtil.getProjectExecutable( project );
-    if( file != null ) {
-      file.delete( true, new SubProgressMonitor( mon, 3 ) );
+    IFile[] files = ResourceUtil.getProjectExecutables( project );
+    for( IFile file: files ) {
+      if( file != null && file.isAccessible() ) {
+        file.delete( true, new SubProgressMonitor( mon, 3 ) );
+      }
     }
   }
